requirements.txt
absl-py==0.7.1
animalai==1.0.3
animalai-train==1.0.3
astor==0.8.0
atari-py==0.2.6
cloudpickle==1.2.1
cycler==0.10.0
dopamine-rl==2.0.5
future==0.17.1
gast==0.2.2
gin-config==0.2.0
grpcio==1.11.1
gym==0.13.1
h5py==2.9.0
jsonpickle==1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.14.5
opencv-python==4.1.0.25
Pillow==5.4.1
protobuf==3.6.1
pyglet==1.3.2
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1.1
scipy==1.3.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.2
termcolor==1.1.0
torch==1.2.0
Werkzeug==0.15.5
